Set up miguelandres' ploopy thumb trackball
This configuration emulates the default configuration for Logitech's MX
Ergos, where the left buttons act as forward (button 5) and back (button
4).

These buttons are the ones closest to the ball, the smaller one being
back.

This is different from the default configuration in which the large
button on the left is back (now forward) and the rightmost button is
forward.

That rightmost button now enables a media control layer, where the back
and forward buttons behave the same ant then the middle button is
play/pause.